+++ samples/fifo/AbstractFifo.cc
@@ -31,8 +31,8 @@ void AbstractFifo::initialize()
     qlenSignal = registerSignal("qlen");
     busySignal = registerSignal("busy");
     queueingTimeSignal = registerSignal("queueingTime");
-    emit(qlenSignal, (long)queue.length());
-    emit(busySignal, (long)0);
+    emit(qlenSignal, queue.length());
+    emit(busySignal, 0);
 }
 
 void AbstractFifo::handleMessage(cMessage *msg)
@@ -43,12 +43,12 @@ void AbstractFifo::handleMessage(cMessage *msg)
         if (queue.empty())
         {
             msgServiced = NULL;
-            emit(busySignal, (long)0);
+            emit(busySignal, 0);
         }
         else
         {
             msgServiced = (cMessage *) queue.pop();
-            emit(qlenSignal, (long)queue.length());
+            emit(qlenSignal, queue.length());
             emit(queueingTimeSignal, simTime() - msgServiced->getTimestamp());
             simtime_t serviceTime = startService( msgServiced );
             scheduleAt( simTime()+serviceTime, endServiceMsg );
@@ -61,14 +61,14 @@ void AbstractFifo::handleMessage(cMessage *msg)
         emit(queueingTimeSignal, 0.0);
         simtime_t serviceTime = startService( msgServiced );
         scheduleAt( simTime()+serviceTime, endServiceMsg );
-        emit(busySignal, (long)1);
+        emit(busySignal, 1);
     }
     else
     {
         arrival( msg );
         queue.insert( msg );
         msg->setTimestamp();
-        emit(qlenSignal, (long)queue.length());
+        emit(qlenSignal, queue.length());
     }
 }
