+++ samples/queueinglib/ResourceBasedQueue.cc
@@ -32,9 +32,9 @@ void ResourceBasedQueue::initialize()
     droppedSignal = registerSignal("dropped");
     queueingTimeSignal = registerSignal("queueingTime");
     queueLengthSignal = registerSignal("queueLength");
-    emit(queueLengthSignal, 0l);
+    emit(queueLengthSignal, 0);
     busySignal = registerSignal("busy");
-    emit(busySignal, 0l);
+    emit(busySignal, 0);
 
     endServiceMsg = new cMessage("end-service");
     fifo = par("fifo");
@@ -59,14 +59,14 @@ void ResourceBasedQueue::handleMessage(cMessage *msg)
         if (!queue.empty() && allocateResource(peek()))
         {
             jobServiced = getFromQueue();
-            emit(queueLengthSignal, (long)length());
+            emit(queueLengthSignal, length());
             simtime_t serviceTime = startService( jobServiced );
             scheduleAt( simTime()+serviceTime, endServiceMsg );
         }
         else
         {
             jobServiced = NULL;
-            emit(busySignal, 0l);
+            emit(busySignal, 0);
         }
     }
     else
@@ -79,7 +79,7 @@ void ResourceBasedQueue::handleMessage(cMessage *msg)
         {
             // processor was idle and the allocation is successful
             jobServiced = job;
-            emit(busySignal, 1l);
+            emit(busySignal, 1);
             simtime_t serviceTime = startService( jobServiced );
             scheduleAt( simTime()+serviceTime, endServiceMsg );
         }
@@ -90,12 +90,12 @@ void ResourceBasedQueue::handleMessage(cMessage *msg)
             {
                 EV << "Capacity full! Job dropped.\n";
                 if (ev.isGUI()) bubble("Dropped!");
-                emit(droppedSignal, 1l);
+                emit(droppedSignal, 1);
                 delete job;
                 return;
             }
             queue.insert( job );
-            emit(queueLengthSignal, (long)length());
+            emit(queueLengthSignal, length());
             job->setQueueCount(job->getQueueCount() + 1);
         }
     }
@@ -189,8 +189,8 @@ void ResourceBasedQueue::resourceGranted(IResourcePool *provider)
     if (!jobServiced && !queue.empty())
     {
         jobServiced = getFromQueue();
-        emit(queueLengthSignal, (long)length());
-        emit(busySignal, 1l);
+        emit(queueLengthSignal, length());
+        emit(busySignal, 1);
         simtime_t serviceTime = startService( jobServiced );
         scheduleAt( simTime()+serviceTime, endServiceMsg );
     }
